api/requirements.txt
databases==0.9.0
fastapi==0.115.12
python-dotenv==1.0.1
SQLAlchemy==2.0.29
asyncpg==0.30.0
black==25.1.0
psycopg2-binary==2.9.10
psycopg2==2.9.10
uvicorn[standard]==0.34.0
gunicorn==23.0.0
pytest==8.3.5
httpx==0.28.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi